refactor: shadowImg 관련 함수들 하나의 유틸 파일로 묶기
