+++ frontend/src/hooks/play/useDragForWeb.ts
@@ -3,8 +3,7 @@ import { usePlayStore } from 'stores/playStore';
 import { getPrevPosition } from 'utils/getPrevPosition';
 
 export const useDragForWeb = (characterInfo: CharacterInfoInterface) => {
-  const { selectedCharacter, setSelectedCharacter, setPrevPosition } =
-    usePlayStore();
+  const { setSelectedCharacter, setPrevPosition } = usePlayStore();
 
   const handleDragStart = (e: React.DragEvent<HTMLImageElement>) => {
     setSelectedCharacter(characterInfo);
